requirements.txt
discord.py
pytz
pandas
pandasql
tabulate
xlrd
pygelbooru